src/third_party/blink/web_tests/fast/events/touch/gesture/gesture-tap-frame-removed-expected.txt

Verifies that a tap occurring on an iframe that gets removed during tap handling doesn't cause a crash.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test case: Remove during mousemove
Sending GestureTap
Received mousemove in child frame
Removing iframe
PASS document.getElementById('target') is null
iframe loaded


Test case: Remove during mousedown
Sending GestureTap
FAIL document.getElementById('target') should be null. Was [object HTMLIFrameElement].
iframe loaded


Test case: Remove during mouseup
Sending GestureTap
FAIL document.getElementById('target') should be null. Was [object HTMLIFrameElement].
iframe loaded


PASS successfullyParsed is true

TEST COMPLETE
